Updated all tests
Deleted all Thread.sleep and Selenium Wait added
