+++ src/test/java/Homework16Test.java
@@ -1,17 +1,13 @@
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import java.time.Duration;
 
 public class Homework16Test extends BaseTest {
     @Test
-    public void registrationNavigation () throws InterruptedException {
+    public void registrationNavigation ()  {
