+++ src/test/resources/TestNG.xml
@@ -7,6 +7,7 @@
         <classes>
             <class name="LoginTests"/>
             <class name="BaseTest"/>
+            <class name="CreateNewPlaylist"/>
             <class name="Homework16Test"/>
             <class name="Homework17Test"/>
             <class name="Homework18Test"/>
